+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2025-02-06
+
+### Changed
+
+- [connector] The Morio connector is now backend by Vector (instead of Logstash)
+- [core] We now use `{|` and `|}` as Mustache delimiters to prevent mixups with VRL delimiters
+- [client] We now use `{|` and `|}` as Mustache delimiters (for consistency with core)
+
 ## [0.6.2] - 2025-02-04
 
 ### Added